services/raven: Extract setup() method
